miniprogram/requirement.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-filter==1.1.0
django-redis==4.8.0
django-rest-swagger==2.1.2
djangorestframework==3.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
mysqlclient==1.3.12
openapi-codec==1.3.2
pytz==2017.2
redis==2.10.6
requests==2.18.4
simplejson==3.11.1
uritemplate==3.0.0
urllib3==1.22
django-extensions==1.9.8
pycrypto==2.6.1
xmltodict==0.11.0
ipython==6.2.1
qiniu==7.2.0
jsonfield==1.0.3
django-cors-headers==2.1.0